requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.1.5
django-polymorphic==3.0.0
djangorestframework==3.12.2
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==2.10
mysqlclient==2.0.3
packaging==20.9
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2020.5
redis==3.5.3
requests==2.25.1
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
